Changed int to size_t for GCC.

+++ src/fof.h
@@ -33,6 +33,6 @@ void fof_search_serial(struct space *s);
 void fof_search_cell(struct space *s, struct cell *c, int *pid, int *num_in_groups, int *num_groups);
 void fof_search_pair_cells(struct space *s, struct cell *ci, struct cell *cj, int *pid, int *num_in_groups, int *num_groups);
 void fof_search_tree_serial(struct space *s);
-void fof_dump_group_data(char *out_file, const int nr_gparts, int *pid, int *num_in_groups);
+void fof_dump_group_data(char *out_file, const size_t nr_gparts, int *pid, int *num_in_groups);
 
 #endif /* SWIFT_FOF_H */
